refactor(worker): extract workers/responses.mjs envelope builders from api.mjs (#510)
Fifth slice of the api.mjs de-monolith — the response-envelope layer. Move
contractVersion, publishedAt, dataResponse, and envelopeResponse into a new leaf
module workers/responses.mjs.

Pure restructuring — NO behaviour change. responses.mjs imports only
CONTRACT_VERSION (../src/contracts.mjs), apiHeaders/weakEtag (./http.mjs), and
latestPointer (./storage.mjs) — all leaves; it calls nothing back into api.mjs,
so no cycle. api.mjs imports the 4 back; dropped the now-unused CONTRACT_VERSION
import (its only remaining users moved into responses.mjs). latestPointer stays
imported in api.mjs (still used directly by handlers).

With config/http/storage/responses/list-query extracted, the whole
request-serving infrastructure layer is now out of api.mjs and the handler
groups can be split next.

api.mjs: 3064 -> 3014 LOC (3570 at the start of #510). None of the 4 are
test-imported (exercised via handleRequest), so no re-exports.

Verified: 229 worker tests pass across 6 suites, worker:test (esbuild) green,
validate:contract-drift clean, lint clean. No logic diff.
